% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A plot of population data for a given country}
\usage{
CountryPopulation(x)
}
\arguments{
\item{x}{A country as a string}
}
\value{
A plot of the population data of that country
}
\description{
This function will take the input of a country and produce a plot of the population data from 1950 to 2022
}
\examples{
CountryPopulation("Italy")
CountryPopulation("Ethiopia")
}
