% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Attendance at the World Cup over time}
\format{
a data frame with 110 observations and 5 columns
\descrive{
\item{Year}{Year the World Cup was held.}
\item{Hosts}{The nation(s) that hosted that World Cup.}
\item{Matches}{The number of matches played in the tournament}
\item{Totatattendence}{The total number of people who attended the World Cup}
\item{Averageattendence}{The average number of people who attended each match}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}?
}
\usage{
World_Cup
}
\description{
This data is from the FIFA World Cups from 1930 t0 2022
}
\keyword{datasets}
