% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population from 1950-2020 by country}
\format{
a data frame with 16920 observations with 72 columns:
\describe {
\item{Year}{The year the population was recorded}
\item{Population}{Total population, both sexes combined, as of 1 July (thousands)}
}
}
\usage{
WorldPopulation
}
\description{
This data was collected by the United Nations in August 2019. The data spans the time of 1950 to 2020.
}
\keyword{datasets}
